requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=1.0.0
langchain-ibm

ibm_watson_machine_learning
langgraph_supervisor
langchain-huggingface
pyodata

sentence-transformers
pdfplumber
langchain-qdrant
fastembed

langsmith

numpy
pandas
scipy
